python: support v5 analytics parameters, refactor v3 delivery metrics
